requirements.txt
flask
google-generativeai
pandas
openai
facebook-scraper
lxml_html_clean
browser_cookie3
google-api-python-client
python-dotenv